Consuming REST API

+++ src/main/java/com/product/api/service/ProductService.java
@@ -1,12 +1,10 @@
 package com.product.api.service;
 
-import java.math.BigInteger;
-
 import com.product.api.model.Products;
 
 public interface ProductService {
 
-	public Products getProduct(BigInteger productId);
+	public Products getProduct(Long productId);
 
 	public void updatePriceByProduct(Products product);
 }
